Sharing callback better defined
