+++ src/Interfaces/IHookParams.php
@@ -2,8 +2,8 @@
 
 namespace Hexbatch\Things\Interfaces;
 
+use Hexbatch\Things\Enums\TypeOfCallback;
 use Hexbatch\Things\Enums\TypeOfHookMode;
-use Hexbatch\Things\Enums\TypeOfHookScope;
 
 interface IHookParams
 {
@@ -15,9 +15,16 @@ interface IHookParams
     public function isHookOn():  bool;
     public function isBlocking():  bool;
     public function isWriting():  bool;
+    public function isSharing():  bool;
     public function getHookMode():  ?TypeOfHookMode;
-    public function getHookScope():  TypeOfHookScope;
     public function  getHookName() :?string;
     public function  getHookNotes() :?string;
+
+    public function getCallbackType():  ?TypeOfCallback;
+    public function getDataTemplate():  array;
+    public function getHeaderTemplate():  array;
+
+    public function  getAddress() :string;
+    public function  getSharedTtl() :?int;
 
 }
