+++ src/OpenApi/Hooks/HookResponse.php
@@ -31,18 +31,17 @@ class HookResponse extends HookParams implements  JsonSerializable
 
         parent::__construct(
             mode: $hook->hook_mode,
-            scope:  $hook->hook_scope,
-            name:  $hook->hook_name,
-            notes:  $hook->hook_notes,
-            action_type:  $hook->getAction()?->getActionType(),
-            action_id:  $hook->getAction()?->getActionId(),
+            name: $hook->hook_name,
+            notes: $hook->hook_notes,
+            action_type: $hook->getAction()?->getActionType(),
+            action_id: $hook->getAction()?->getActionId(),
             constant_data: $hook->hook_constant_data?->getArrayCopy()??[],
             tags: $hook->hook_tags?->getArrayCopy()??[],
             hook_on: $hook->is_on,
             is_blocking: $hook->is_blocking,
             is_writing: $hook->is_writing_data_to_thing,
+            is_sharing: $hook->is_sharing,
             callback_type: $hook->hook_callback_type,
-            sharing: $hook->hook_sharing_type,
             ttl_shared: $hook->ttl_shared,
             data_template: $hook->hook_data_template?->getArrayCopy()??[],
             header_template: $hook->hook_header_template?->getArrayCopy()??[],
